Refactor ConvertJsonToCompaniesUseCase.kt
