+++ app/src/main/java/com/harshnandwani/digitaltijori/domain/use_case/company/ConvertJsonToCompaniesUseCase.kt
@@ -1,10 +1,10 @@
-package com.harshnandwani.digitaltijori.domain.use_case
+package com.harshnandwani.digitaltijori.domain.use_case.company
 
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.harshnandwani.digitaltijori.domain.model.Company
 
-class ConvertJsonToCompanies {
+class ConvertJsonToCompaniesUseCase {
 
     operator fun invoke(companiesJson: String): List<Company> {
         val typeToken = object : TypeToken<List<Company>>() {}.type
